Add adaptor unit tests
This commit adds unit tests of the boltdb adaptor to the package.
The tests use a temporary database files and boltdb.MessageHandler
interface mock.

bucket utility function was modified to return "bucket not found"
errors when the passed transaction is read-only.

A thin boltdb wrapper TempDB was introduced to the boltdb/internal
package to facilitate tests set-ups.